requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
requests==2.32.3
pydantic==2.9.1
starlette==0.38.2
numpy==1.26.4
python-multipart==0.0.20
python-dotenv==1.1.1
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
psycopg2-binary==2.9.6